commit at ${COMMIT_TIMESTAMP}

+++ public/js/property.js
@@ -130,7 +130,7 @@ function propertyCard(property, isAuth, imageEventListner, isOwn = false) {
     favoritBtn.appendChild(materialSymbolsOutlined3);
 
     const commentBtn = document.createElement('a');
-    commentBtn.href = "/property/" + property._id+"#q&a";
+    commentBtn.href = "/property/" + property._id + "#q&a";
     commentBtn.classList.add("comment-btn");
 
     const materialSymbolsOutlined4 = document.createElement('span');
@@ -239,7 +239,7 @@ function propertyCard(property, isAuth, imageEventListner, isOwn = false) {
         const editBtn = document.createElement('button');
         editBtn.classList.add("ownBtn");
         editBtn.classList.add("btn1");
-        editBtn.type="reset"
+        editBtn.type = "reset"
         editBtn.appendChild(document.createTextNode("EDIT"))
         const deleteBtn = document.createElement('button');
         deleteBtn.classList.add("ownBtn");
@@ -252,28 +252,32 @@ function propertyCard(property, isAuth, imageEventListner, isOwn = false) {
 
         // delete btn
         deleteBtn.addEventListener("click", () => {
-            fetch(host + "/admin/delete-property", {
-                    method: 'DELETE',
-                    body: new URLSearchParams("propId=" + property._id),
-                })
-                .then(result => result.json())
-                .then(result => {
-                    if (result.statusCode == 200) {
-                        div.remove();
-                    } else {
-                        alert(result.message)
-                    }
-                })
-                .catch(err => {
-                    alert("property is not deleted")
-                    console.log(err)
-                });
+            var result = confirm("Want to delete?");
+            if (result) {
+                fetch(host + "/admin/delete-property", {
+                        method: 'DELETE',
+                        body: new URLSearchParams("propId=" + property._id),
+                    })
+                    .then(result => result.json())
+                    .then(result => {
+                        if (result.statusCode == 200) {
+                            div.remove();
+                        } else {
+                            alert(result.message)
+                        }
+                    })
+                    .catch(err => {
+                        alert("property is not deleted")
+                        console.log(err)
+                    });
+            }
+
         });
 
         // edit btn
-        editBtn.addEventListener("click",()=>{
-            console.log(host+"/edit-property/"+property._id);
-            window.location.href=host+"/admin/edit-property/"+property._id;
+        editBtn.addEventListener("click", () => {
+            console.log(host + "/edit-property/" + property._id);
+            window.location.href = host + "/admin/edit-property/" + property._id;
         })
     }
